requirements.txt
pedalboard>=0.9.0
aalink>=0.1.0
python-rtmidi>=1.5.0
mido>=1.3.0
numpy>=1.24.0
sounddevice>=0.5.0
